+++ CHANGELOG.md
@@ -7,6 +7,8 @@ _Unreleased_
 * Failsafe for tapping "back" button in the editor and losing changes.
 * Press "back" on the main screen to return to your location.
 * When adding a new opening hours fragment, enter the time interval.
+* Hours fragments are sorted and de-duplicated on save.
+* Current floor is displayed in the editor even without an address.
 * Fixed certificate error and icon for Android 7.
 
 ## 0.1.2